chore(popup): direction-value型のヘルパー関数をutilsに移動
