+++ packages/wiz-ui-react/src/components/base/popup/types/direction.ts
@@ -16,18 +16,3 @@ export const DIRECTION_MAP = {
 export type DirectionKey = keyof typeof DIRECTION_MAP;
 
 export type DirectionValue = (typeof DIRECTION_MAP)[DirectionKey];
-
-export const createDirectionValue = ({
-  first,
-  second,
-}:
-  | {
-      first: "b" | "t";
-      second: "l" | "c" | "r";
-    }
-  | {
-      first: "l" | "r";
-      second: "t" | "c" | "b";
-    }) => {
-  return `${first}${second}` as DirectionValue;
-};
