+++ packages/wiz-ui-react/src/components/base/popup/utils/direction-value.ts
@@ -1,0 +1,16 @@
+import { DirectionValue } from "../types/direction";
+
+export const createDirectionValue = ({
+  first,
+  second,
+}:
+  | {
+      first: "b" | "t";
+      second: "l" | "c" | "r";
+    }
+  | {
+      first: "l" | "r";
+      second: "t" | "c" | "b";
+    }) => {
+  return `${first}${second}` as DirectionValue;
+};
